+++ app/search/search_test.js
@@ -191,6 +191,18 @@
                 delete objToFilter.statuses;
                 expect(ctrl.certificationStatusFilter(objToFilter)).toBe(true);
             });
+
+            it('should have a function to get the right icon for a status', function () {
+                expect(ctrl.statusFont).toBeDefined();
+            });
+
+            it('should get the right icon for various statuses', function () {
+                expect(ctrl.statusFont('Active')).toBe('fa-check-circle status-good');
+                expect(ctrl.statusFont('Suspended')).toBe('fa-check-circle status-good');
+                expect(ctrl.statusFont('Retired')).toBe('fa-close status-bad');
+                expect(ctrl.statusFont('Terminated')).toBe('fa-close status-bad');
+                expect(ctrl.statusFont('Withdrawn')).toBe('fa-close status-bad');
+            });
         });
     });
 })();